完善角色管理的权限校验机制

+++ Model-Web/src/main/java/com/cn/lucky/morning/model/web/controller/admin/RoleController.java
@@ -7,11 +7,13 @@ import com.cn.lucky.morning.model.common.mvc.MvcResult;
 import com.cn.lucky.morning.model.domain.Authority;
 import com.cn.lucky.morning.model.domain.AuthorityGroup;
 import com.cn.lucky.morning.model.domain.Role;
+import com.cn.lucky.morning.model.domain.User;
 import com.cn.lucky.morning.model.domain.customer.RoleEx;
 import com.cn.lucky.morning.model.service.AuthorityService;
 import com.cn.lucky.morning.model.service.RoleService;
 import com.google.common.base.Objects;
 import org.apache.commons.lang.StringUtils;
+import org.apache.shiro.SecurityUtils;
 import org.apache.shiro.authz.annotation.Logical;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.springframework.stereotype.Controller;
@@ -58,8 +60,8 @@ public class RoleController {
     @RequestMapping("/add")
     @RequiresPermissions(value = {"ROLE_ADD", Const.role.ROLE_SUPER}, logical = Logical.OR)
     public String add(Model model) {
-        Map<AuthorityGroup,List<Authority>> authorities = authorityService.getAuthority();
-        model.addAttribute("authorities",authorities);
+        Map<AuthorityGroup, List<Authority>> authorities = authorityService.getAuthority();
+        model.addAttribute("authorities", authorities);
         return "admin/role/add";
     }
 
@@ -73,6 +75,11 @@ public class RoleController {
                 result.setSuccess(false);
                 result.setMessage("添加失败：角色名称不能为空");
             } else {
+                User user = (User) SecurityUtils.getSubject().getPrincipal();
+                Role userRole = roleService.getById(user.getRoleId());
+                if (!Objects.equal(userRole.getIsSuper(), Const.role.IS_SUPER)) {
+                    role.setIsSuper(Const.role.NOT_SUPER);
+                }
                 role.setAuthority(StringUtils.join(role.getAuthoritys(), ','));
                 boolean success = roleService.add(role);
                 if (!success) {
@@ -90,8 +97,8 @@ public class RoleController {
     @RequestMapping("/edit")
     @RequiresPermissions(value = {"ROLE_UPDATE", Const.role.ROLE_SUPER}, logical = Logical.OR)
     public String edit(Long id, Model model) {
-        Map<AuthorityGroup,List<Authority>> authorities = authorityService.getAuthority();
-        model.addAttribute("authorities",authorities);
+        Map<AuthorityGroup, List<Authority>> authorities = authorityService.getAuthority();
+        model.addAttribute("authorities", authorities);
         Role role = roleService.getById(id);
         if (!StringUtils.isEmpty(role.getAuthority())) {
             model.addAttribute("roleAuth", Arrays.asList(role.getAuthority().split(",")));
@@ -110,6 +117,11 @@ public class RoleController {
                 result.setSuccess(false);
                 result.setMessage("修改失败：角色名称不能为空");
             } else {
+                User user = (User) SecurityUtils.getSubject().getPrincipal();
+                Role userRole = roleService.getById(user.getRoleId());
+                if (!Objects.equal(userRole.getIsSuper(), Const.role.IS_SUPER)) {
+                    role.setIsSuper(Const.role.NOT_SUPER);
+                }
                 if (!Objects.equal(role.getIsSuper(), Const.role.IS_SUPER)) {
                     role.setAuthority(StringUtils.join(role.getAuthoritys(), ','));
                 }
@@ -132,10 +144,21 @@ public class RoleController {
     public MvcResult delete(Long id) {
         MvcResult result = MvcResult.create();
         try {
-            boolean success = roleService.delete(id);
-            if (!success) {
+            User user = (User) SecurityUtils.getSubject().getPrincipal();
+            Role role = roleService.getById(user.getRoleId());
+            Role delRole = roleService.getById(id);
+            if (Objects.equal(id, role.getId())) {
                 result.setSuccess(false);
-                result.setMessage("删除失败：未知原因");
+                result.setMessage("不能删除自己的角色");
+            } else if (delRole != null && Objects.equal(delRole.getIsSuper(), Const.role.IS_SUPER) && !Objects.equal(role.getIsSuper(), Const.role.IS_SUPER)) {
+                result.setSuccess(false);
+                result.setMessage("无权限删除超级管理员");
+            } else {
+                boolean success = roleService.delete(id);
+                if (!success) {
+                    result.setSuccess(false);
+                    result.setMessage("删除失败：未知原因");
+                }
             }
         } catch (Exception e) {
             result.setSuccess(false);
@@ -150,10 +173,26 @@ public class RoleController {
     public MvcResult deleteList(Long[] ids) {
         MvcResult result = MvcResult.create();
         try {
-            boolean success = roleService.deleteList(Arrays.asList(ids));
-            if (!success) {
-                result.setSuccess(false);
-                result.setMessage("删除失败：未知原因");
+            User user = (User) SecurityUtils.getSubject().getPrincipal();
+            Role role = roleService.getById(user.getRoleId());
+            for (Long id : ids) {
+                Role delRole = roleService.getById(id);
+                if (Objects.equal(id, role.getId())) {
+                    result.setSuccess(false);
+                    result.setMessage("不能删除自己的角色");
+                    break;
+                } else if (delRole != null && Objects.equal(delRole.getIsSuper(), Const.role.IS_SUPER) && !Objects.equal(role.getIsSuper(), Const.role.IS_SUPER)) {
+                    result.setSuccess(false);
+                    result.setMessage("无权限删除超级管理员");
+                    break;
+                }
+            }
+            if (result.isSuccess()) {
+                boolean success = roleService.deleteList(Arrays.asList(ids));
+                if (!success) {
+                    result.setSuccess(false);
+                    result.setMessage("删除失败：未知原因");
+                }
             }
         } catch (Exception e) {
             result.setSuccess(false);
